+++ temba/campaigns/views.py
@@ -551,10 +551,10 @@ class CampaignEventCRUDL(SmartCRUDL):
             if self.has_org_perm("campaigns.campaignevent_delete"):
                 links.append(
                     dict(
+                        id="event-delete",
                         title="Delete",
-                        delete=True,
-                        success_url=reverse("campaigns.campaign_read", args=[campaign_event.campaign.pk]),
                         href=reverse("campaigns.campaignevent_delete", args=[campaign_event.id]),
+                        modax=_("Delete Event"),
                     )
                 )
 
@@ -563,6 +563,8 @@ class CampaignEventCRUDL(SmartCRUDL):
     class Delete(ModalMixin, OrgObjPermsMixin, SmartDeleteView):
 
         default_template = "smartmin/delete_confirm.html"
+        submit_button_name = _("Delete")
+        fields = ("uuid",)
 
         def get_object_org(self):
             return self.get_object().campaign.org