+++ temba/triggers/views.py
@@ -5,8 +5,6 @@ from django import forms
 from django.db.models import Min
 from django.http import HttpResponse, HttpResponseRedirect
 from django.urls import reverse
-from django.utils import timezone
-from django.utils.timezone import get_current_timezone_name
 from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.models import Channel
@@ -27,7 +25,7 @@ from temba.utils.fields import (
     SelectMultipleWidget,
     SelectWidget,
 )
-from temba.utils.views import BulkActionMixin
+from temba.utils.views import BulkActionMixin, ComponentFormMixin
 
 from .models import Trigger
 
@@ -37,12 +35,7 @@ class BaseTriggerForm(forms.ModelForm):
     Base form for creating different trigger types
     """
 
-    flow = forms.ModelChoiceField(
-        Flow.objects.filter(pk__lt=0),
-        label=_("Flow"),
-        required=True,
-        widget=SelectWidget(attrs={"widget_only": True, "placeholder": "Select the flow to trigger"}),
-    )
+    flow = forms.ModelChoiceField(Flow.objects.filter(pk__lt=0), label=_("Flow"), required=True,)
 
     def __init__(self, user, flows, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -236,11 +229,20 @@ class ScheduleTriggerForm(BaseScheduleForm, forms.ModelForm):
     repeat_period = forms.ChoiceField(
         choices=Schedule.REPEAT_CHOICES, label="Repeat", required=False, widget=SelectWidget()
     )
-    repeat_days_of_week = forms.CharField(required=False)
-    start = forms.ChoiceField(
-        choices=(("stop", "Stop Schedule"), ("later", "Schedule for later")), widget=SelectWidget()
+
+    repeat_days_of_week = forms.MultipleChoiceField(
+        choices=Schedule.REPEAT_DAYS_CHOICES,
+        label="Repeat Days",
+        required=False,
+        widget=SelectMultipleWidget(attrs=({"placeholder": _("Select days to repeat on")})),
     )
-    start_datetime_value = forms.IntegerField(required=False)
+
+    start_datetime = forms.DateTimeField(
+        required=False,
+        label=_("Start Time"),
+        widget=InputWidget(attrs={"datetimepicker": True, "placeholder": "Select a time to start the flow"}),
+    )
+
     flow = forms.ModelChoiceField(
         Flow.objects.filter(pk__lt=0),
         label=_("Flow"),
@@ -261,16 +263,21 @@ class ScheduleTriggerForm(BaseScheduleForm, forms.ModelForm):
     def __init__(self, user, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.user = user
-        flows = Flow.get_triggerable_flows(user.get_org())
+        org = user.get_org()
+        flows = Flow.get_triggerable_flows(org)
 
+        self.fields["start_datetime"].help_text = _("%s Time Zone" % org.timezone)
         self.fields["flow"].queryset = flows
+
+    def clean_repeat_days_of_week(self):
+        return "".join(self.cleaned_data["repeat_days_of_week"])
 
     def clean_omnibox(self):
         return omnibox_deserialize(self.user.get_org(), self.cleaned_data["omnibox"])
 
     class Meta:
         model = Trigger
-        fields = ("flow", "omnibox", "repeat_period", "repeat_days_of_week", "start", "start_datetime_value")
+        fields = ("flow", "omnibox", "repeat_period", "repeat_days_of_week", "start_datetime")
 
 
 class InboundCallTriggerForm(GroupBasedTriggerForm):
@@ -414,7 +421,7 @@ class TriggerCRUDL(SmartCRUDL):
 
             add_section("trigger-catchall", "triggers.trigger_catchall", "icon-bubble")
 
-    class Update(ModalMixin, OrgMixin, SmartUpdateView):
+    class Update(ModalMixin, ComponentFormMixin, OrgMixin, SmartUpdateView):
         success_message = ""
         trigger_forms = {
             Trigger.TYPE_KEYWORD: KeywordTriggerForm,
@@ -434,8 +441,6 @@ class TriggerCRUDL(SmartCRUDL):
             context = super().get_context_data(**kwargs)
             if self.get_object().schedule:
                 context["days"] = self.get_object().schedule.repeat_days_of_week or ""
-            context["user_tz"] = get_current_timezone_name()
-            context["user_tz_offset"] = int(timezone.localtime(timezone.now()).utcoffset().total_seconds() // 60)
             return context
 
         def form_invalid(self, form):
@@ -452,7 +457,18 @@ class TriggerCRUDL(SmartCRUDL):
             if trigger_type == Trigger.TYPE_SCHEDULE:
                 repeat_period = trigger.schedule.repeat_period
                 omnibox = omnibox_serialize(trigger.org, trigger.groups.all(), trigger.contacts.all())
-                return dict(repeat_period=repeat_period, omnibox=omnibox)
+
+                repeat_days_of_week = []
+                if trigger.schedule.repeat_days_of_week:
+                    repeat_days_of_week = list(trigger.schedule.repeat_days_of_week)
+
+                return dict(
+                    repeat_period=repeat_period,
+                    omnibox=omnibox,
+                    start_datetime=trigger.schedule.next_fire,
+                    repeat_days_of_week=repeat_days_of_week,
+                )
+            return super().derive_initial()
 
         def get_form_kwargs(self):
             kwargs = super().get_form_kwargs()
@@ -466,9 +482,11 @@ class TriggerCRUDL(SmartCRUDL):
             if trigger_type == Trigger.TYPE_SCHEDULE:
                 schedule = trigger.schedule
 
+                form.cleaned_data.get("repeat_days_of_week")
+
                 # update our schedule
                 schedule.update_schedule(
-                    form.get_start_time(schedule.org.timezone),
+                    form.cleaned_data.get("start_datetime"),
                     form.cleaned_data.get("repeat_period"),
                     form.cleaned_data.get("repeat_days_of_week"),
                 )
@@ -553,7 +571,7 @@ class TriggerCRUDL(SmartCRUDL):
         def get_queryset(self, *args, **kwargs):
             return super().get_queryset(*args, **kwargs).filter(is_active=True, is_archived=True)
 
-    class CreateTrigger(OrgPermsMixin, SmartCreateView):
+    class CreateTrigger(OrgPermsMixin, ComponentFormMixin, SmartCreateView):
         success_url = "@triggers.trigger_list"
         success_message = ""
 
@@ -646,8 +664,6 @@ class TriggerCRUDL(SmartCRUDL):
 
         def get_context_data(self, **kwargs):
             context = super().get_context_data(**kwargs)
-            context["user_tz"] = get_current_timezone_name()
-            context["user_tz_offset"] = int(timezone.localtime(timezone.now()).utcoffset().total_seconds() // 60)
             return context
 
         def form_invalid(self, form):
@@ -661,7 +677,7 @@ class TriggerCRUDL(SmartCRUDL):
         def form_valid(self, form):
             analytics.track(self.request.user.username, "temba.trigger_created", dict(type="schedule"))
             org = self.request.user.get_org()
-            start_time = form.get_start_time(org.timezone)
+            start_time = form.cleaned_data["start_datetime"]
 
             schedule = Schedule.create_schedule(
                 org,